Fixed 32 times obstacle bug works like charm
