+++ hotaru/hotaru_planner_node_teb/CMakeLists.txt
@@ -142,7 +142,7 @@ include_directories(
 ## Declare a C++ executable
 ## With catkin_make all packages are built within a single CMake context
 ## The recommended prefix ensures that target names across packages don't collide
-add_executable(${PROJECT_NAME}_node src/hotaru_teb_local_planner_node.cpp)
+add_executable(${PROJECT_NAME}_node src/hotaru_teb_local_planner_node.cpp src/hotaru_teb_planner.cpp)
 
 ## Rename C++ executable without prefix
 ## The above recommended prefix causes long target names, the following renames the
