+++ hotaru/hotaru_planner_node_teb/include/hotaru_planner_node_teb/hotaru_teb_local_planner.hpp
@@ -28,6 +28,7 @@ namespace hotaru
 class HotaruTebLocalPlannerNode: public HotaruPlannerNodeRos
 {
 private:
+	int obstacle_waypoint;
 protected:
 	std::vector<teb_local_planner::ObstaclePtr> obst_vector;
 	teb_local_planner::TebConfig tebconfig;
@@ -39,6 +40,7 @@ protected:
 	teb_local_planner::PoseSE2 se2_plan_end;
 	std::vector<teb_local_planner::TrajectoryPointMsg> trajectory_profile;
 	std::shared_ptr<LocalPlannerStateMachine> sm_behav_planner;
+	std::shared_ptr<rei::RosCommunicationGraphNotifier> sm_behav_notifier;
 	std::shared_ptr<SyncGuardLocalPlanner> guard_local_planner;
 	// Select a trajectory slicing method
 	std::unique_ptr<SpeedLookaheadSlicer> trajectoryslicestrategy;
@@ -58,12 +60,12 @@ protected:
 
 	virtual void startAfterSynchronization()
 	{
-		sm_behav_planner->start();
 		timer_planner = nh->createTimer(ros::Duration(1.0/10.0), &HotaruTebLocalPlannerNode::plannerCycle, this);
 		timer_planner.start();
 		timer_trajectory_publish = nh->createTimer(ros::Duration(1.0/20.0), &HotaruTebLocalPlannerNode::cbTrajectoryPublishTimer, this);
 		timer_trajectory_publish.start();
 		timer_obstacle_update = nh->createTimer(ros::Duration(1.0/15.0), &HotaruTebLocalPlannerNode::cbObstacleTimer, this);
+		sm_behav_planner->start();
 		ROS_INFO("START PLANNER AFTER SYNC");
 	}
 
@@ -79,7 +81,7 @@ public:
 			std::shared_ptr<ros::NodeHandle> private_nh,
 			std::string base_frame, std::string target_frame,
 			const bool debug=false):
-		HotaruPlannerNodeRos(nh, private_nh, base_frame, target_frame){}
+		HotaruPlannerNodeRos(nh, private_nh, base_frame, target_frame), obstacle_waypoint(0){}
 
 	virtual void plannerCycle(const ros::TimerEvent& e)
 	{
@@ -95,172 +97,132 @@ public:
 		se2_current_pose = teb_local_planner::PoseSE2(p0);
 	}
 
+	static void originalTrajectory(
+			const hotaru_msgs::RefinedTrajectory& input_trajectory,
+			hotaru_msgs::RefinedTrajectory& refined_trajectory,
+			const int& start_point, const int& end_point)
+	{
+		for (int i = start_point; i < end_point; i++)
+		{
+			refined_trajectory.waypoints.emplace_back(
+				input_trajectory.waypoints[i]
+			);
+		}
+	}
+
+	void publishTrajectory()
+	{
+		pubsubstate->msg_port_refined_trajectory.header.stamp = ros::Time::now();
+		this->port_refined_trajectory.publish(pubsubstate->msg_port_refined_trajectory);
+	}
+
 	void cbTrajectoryPublishTimer(const ros::TimerEvent& e)
 	{
 		pubsubstate->msg_port_refined_trajectory.waypoints.clear();
-		mtx_trajectory_update.lock();
 		// Don't publish anything until no valid trajectory is received
-		if (trajectory_profile.size() > 0)
+		if (sm_behav_planner->isReplanning())
 		{
-			for (int i = 0; i < trajectory_profile.size(); i++)
+			if (trajectory_profile.size() > 0)
 			{
-				hotaru_msgs::Waypoint w0;
-				tf2::doTransform(trajectory_profile[i].pose, w0.pose.pose, inv_transform_base);
-				//w0.twist.twist = v.velocity;
-				w0.twist.twist.linear.x = 5.0;
-				pubsubstate->msg_port_refined_trajectory.waypoints.emplace_back(w0);
-			}
-			mtx_trajectory_update.unlock();
-			//
-			for (int i = trajectoryslicestrategy->getSlicePoint();
-				i < std::min(trajectoryslicestrategy->getSlicePoint() +
+				mtx_trajectory_update.lock();
+				for (int i = 0; i < trajectory_profile.size(); i++)
+				{
+					hotaru_msgs::Waypoint w0;
+					tf2::doTransform(trajectory_profile[i].pose, w0.pose.pose, inv_transform_base);
+					//w0.twist.twist = v.velocity;
+					w0.twist.twist.linear.x = 5.0;
+					pubsubstate->msg_port_refined_trajectory.waypoints.emplace_back(w0);
+				}
+				mtx_trajectory_update.unlock();
+				//
+				originalTrajectory(
+						pubsubstate->msg_port_input_trajectory,
+						pubsubstate->msg_port_refined_trajectory,
+						trajectoryslicestrategy->getSlicePoint(),
+						std::min(trajectoryslicestrategy->getSlicePoint() +
 						hotaru::TrajectoryLookaheadPoint::lookaheadPoint(
 							10.0, pubsubstate->msg_port_current_velocity.twist),
-							static_cast<int>(pubsubstate->msg_port_input_trajectory.waypoints.size())); i++)
-			{
-				pubsubstate->msg_port_refined_trajectory.waypoints.emplace_back(
-					pubsubstate->msg_port_input_trajectory.waypoints[i]
+							static_cast<int>(pubsubstate->msg_port_input_trajectory.waypoints.size()))
 				);
+				publishTrajectory();
 			}
-
-			pubsubstate->msg_port_refined_trajectory.header.stamp = ros::Time::now();
-			this->port_refined_trajectory.publish(pubsubstate->msg_port_refined_trajectory);
 		}
+		else if (sm_behav_planner->isRelay())
+		{
+			int d = std::min(static_cast<int>(pubsubstate->msg_port_input_trajectory.waypoints.size()),
+					std::max(hotaru::TrajectoryLookaheadPoint::lookaheadPoint(10, pubsubstate->msg_port_current_velocity.twist), 40)
+			);
+			originalTrajectory(
+					pubsubstate->msg_port_input_trajectory,
+					pubsubstate->msg_port_refined_trajectory,
+					pubsubstate->msg_port_closests_waypoint.data,
+					pubsubstate->msg_port_closests_waypoint.data + d
+			);
+
+			publishTrajectory();
+		}
+
 	}
 
-	virtual void config()
-	{
-		tebconfig.map_frame = tf_planner_state.getBaseFrame();
-		//
-		if (!private_nh->getParam("dt_ref", tebconfig.trajectory.dt_ref))
-		{
-			tebconfig.trajectory.dt_ref = hotaru::teb::DEFAULT_TEB_CONFIG_TRAJECTORY_DT_REF;
-		}
-		ROS_INFO_STREAM("Using dt ref: " << tebconfig.trajectory.dt_ref);
-		//
-		if (!private_nh->getParam("dt_hysteresis", tebconfig.trajectory.dt_hysteresis))
-		{
-			tebconfig.trajectory.dt_hysteresis = hotaru::teb::DEFAULT_TEB_CONFIG_TRAJECTORY_DT_HYSTERESIS;
-		}
-		ROS_INFO_STREAM("Using dt hysteresis: " << tebconfig.trajectory.dt_hysteresis);
-		//
-		if (!private_nh->getParam("min_turning_radius", tebconfig.robot.min_turning_radius))
-		{
-			tebconfig.robot.min_turning_radius = hotaru::teb::DEFAULT_MIN_TURNING_RADIUS;
-		}
-		ROS_INFO_STREAM("Using minimum turning radius: " << tebconfig.robot.min_turning_radius);
-		tebconfig.robot.max_vel_y = 0.0;
-		tebconfig.robot.max_vel_x_backwards = 0.0;
-		//
-		tebconfig.robot.acc_lim_theta = 0.05;
-		tebconfig.robot.acc_lim_x = 0.2;
-		tebconfig.robot.max_vel_x_backwards = 0.01;
-		if (!private_nh->getParam("wheelbase", tebconfig.robot.wheelbase))
-		{
-			tebconfig.robot.wheelbase = hotaru::teb::DEFAULT_WHEELBASE;
-		}
-		tebconfig.robot.cmd_angle_instead_rotvel = true;
-		// Set minimal obstacle distance
-		if (!private_nh->getParam("min_obstacle_distance", tebconfig.obstacles.min_obstacle_dist))
-		{
-			tebconfig.obstacles.min_obstacle_dist = hotaru::teb::DEFAULT_MIN_OBSTACLE_DISTANCE;
-		}
-		ROS_INFO_STREAM("Using minimum obstacle distance: " << tebconfig.obstacles.min_obstacle_dist);
-		// Set minimal inflation distance
-		if (!private_nh->getParam("inflation_distance", tebconfig.obstacles.min_obstacle_dist))
-		{
-			tebconfig.obstacles.inflation_dist = hotaru::teb::DEFAULT_MIN_INFLATION;
-		}
-		ROS_INFO_STREAM("Using inflation distance: " << tebconfig.obstacles.inflation_dist);
-		// Set minimal weight obstacle
-		if (!private_nh->getParam("weight_obstacle", tebconfig.optim.weight_obstacle))
-		{
-			tebconfig.optim.weight_obstacle = hotaru::teb::DEFAULT_WEIGHT_OBSTACLE;
-		}
-		ROS_INFO_STREAM("Weight obstacle: " << tebconfig.optim.weight_obstacle);
-		//
-		if (!private_nh->getParam("weight_viapoints", tebconfig.optim.weight_viapoint))
-		{
-			tebconfig.optim.weight_viapoint = hotaru::teb::DEFAULT_WEIGHT_VIAPOINTS;
-		}
-		ROS_INFO_STREAM("Using weight viapoints: " << tebconfig.optim.weight_viapoint);
-		//
-		tebconfig.optim.weight_max_vel_y = 0;
-		// Load: weight kinematics turning radius
-		if (!private_nh->getParam("weight_kinematics_turning_radius", tebconfig.optim.weight_kinematics_turning_radius))
-		{
-			tebconfig.optim.weight_kinematics_turning_radius = hotaru::teb::DEFAULT_WEIGHT_KINEMATICS_TURNING_RADIUS;
-		}
-		ROS_INFO_STREAM("Using weight turning radius (kinematics): " << tebconfig.optim.weight_viapoint);
-		// Load: weight forward drive
-		if (!private_nh->getParam("weight_kinematics_forward_drive", tebconfig.optim.weight_kinematics_forward_drive))
-		{
-			tebconfig.optim.weight_kinematics_forward_drive = hotaru::teb::DEFAULT_WEIGHT_KINEMATICS_TURNING_RADIUS;
-		}
-		ROS_INFO_STREAM("Using weight forward drive: " << tebconfig.optim.weight_kinematics_forward_drive);
-		// Load: nh kinematics
-		if (!private_nh->getParam("weight_kinematics_nh", tebconfig.optim.weight_kinematics_nh))
-		{
-			tebconfig.optim.weight_kinematics_nh = hotaru::teb::DEFAULT_WEIGHT_KINEMATICS_NH;
-		}
-		ROS_INFO_STREAM("Using weight kinematics nh: " << tebconfig.optim.weight_kinematics_nh);
-		// Load: xy goal tolerance
-		if (!private_nh->getParam("xy_goal_tolerance", tebconfig.goal_tolerance.xy_goal_tolerance))
-		{
-			tebconfig.goal_tolerance.xy_goal_tolerance = hotaru::teb::DEFAULT_XY_GOAL_TOLERANCE;
-		}
-		ROS_INFO_STREAM("Using xy goal tolerance: " << tebconfig.goal_tolerance.xy_goal_tolerance);
-		// Load: yaw goal tolerance
-		if (!private_nh->getParam("yaw_goal_tolerance", tebconfig.goal_tolerance.yaw_goal_tolerance))
-		{
-			tebconfig.goal_tolerance.yaw_goal_tolerance = hotaru::teb::DEFAULT_YAW_GOAL_TOLERANCE;
-		}
-		ROS_INFO_STREAM("Using yaw goal tolerance: " << tebconfig.goal_tolerance.yaw_goal_tolerance);
-	}
+	virtual void config();
 
-	virtual void executeUpdate_velocity() override
+
+	virtual void executeUpdate_velocity(const geometry_msgs::TwistStamped::ConstPtr& msg) override
 	{
 		const double speed = sqrt(
-				(pubsubstate->msg_port_current_velocity.twist.linear.x*pubsubstate->msg_port_current_velocity.twist.linear.x)+
-				(pubsubstate->msg_port_current_velocity.twist.linear.y*pubsubstate->msg_port_current_velocity.twist.linear.y));
+				(msg->twist.linear.x*pubsubstate->msg_port_current_velocity.twist.linear.x)+
+				(msg->twist.linear.y*pubsubstate->msg_port_current_velocity.twist.linear.y));
 		trajectoryslicestrategy->setSpeed(speed);
 	}
 
-	virtual void execute_update_input_trajectory() override
+	virtual void execute_update_input_trajectory(const hotaru_msgs::RefinedTrajectory::ConstPtr& msg) override
 	{
 
 	}
 
-	virtual void executeUpdate_closest_waypoint() override
+	virtual void executeUpdate_closest_waypoint(const std_msgs::Int32::ConstPtr& msg) override
 	{
-		trajectoryslicestrategy->setClosestWaypointIndex(pubsubstate->msg_port_closests_waypoint.data);
+		trajectoryslicestrategy->setClosestWaypointIndex(msg->data);
+		if (sm_behav_planner->isReplanning())
+		{
+			if (msg->data > obstacle_waypoint || msg->data < 5)
+			{
+				sm_behav_planner->propagateSignal(
+					std::make_shared<hotaru::trajectory_signals::SignalNoObstacleDetected>(
+							ros::Time::now().toNSec())
+				);
+				sm_behav_planner->stepstatemachine();
+			}
+		}
 	}
 
-	virtual void executeReplan_request_sig() override
+	virtual void executeReplan_request_sig(const rei_planner_signals::ReplanRequest::ConstPtr& msg) override
 	{
-		if (pubsubstate->msg_port_replan_request_sig.eval)
+		if (msg->eval)
 		{
 			sm_behav_planner->propagateSignal(
 				std::make_shared<hotaru::trajectory_signals::SignalReplanningTrajectory>(
-						pubsubstate->msg_port_replan_request_sig.header.stamp.toNSec())
+						msg->header.stamp.toNSec())
 			);
 			sm_behav_planner->stepstatemachine();
 		}
 
 	}
 
-	virtual void execute_update_current_pose() override
+	virtual void execute_update_current_pose(const geometry_msgs::PoseStamped::ConstPtr& msg) override
 	{
 
 	}
 
 	void onReplan()
 	{
+		planner->clearPlanner();
 		ROS_INFO("RUNAWAY");
 	}
 
 	void onStart()
 	{
+		ROS_INFO("Started behavior state machine");
 
 	}
 
@@ -271,7 +233,8 @@ public:
 
 	void onRelay()
 	{
-
+		obstacle_waypoint = 0;
+		ROS_INFO("RELAY");
 	}
 
 	void cbObstacleTimer(const ros::TimerEvent& e)
@@ -317,9 +280,10 @@ public:
 		return true;
 	}
 
-	virtual void executeUpdate_obstacles() override
+	virtual void executeUpdate_obstacles(const rei_monitoring_msgs::DetectedObstacles::ConstPtr& msg) override
 	{
-
+		obstacle_waypoint = msg->max_closest_waypoint +
+				std::max(TrajectoryLookaheadPoint::lookaheadPoint(4, pubsubstate->msg_port_current_velocity.twist), 20);
 	}
 
 	virtual bool assignSyncGuards() override
@@ -345,7 +309,10 @@ public:
 		setSyncStateMachineCallbacks();
 		//  Setup planner SM callbacks
 		guard_local_planner = std::make_shared<SyncGuardLocalPlanner>(sync_state_machine);
-		sm_behav_planner = std::make_shared<LocalPlannerStateMachine>(notifier, guard_local_planner);
+		sm_behav_notifier = std::make_shared<rei::RosCommunicationGraphNotifier>("hotaru_planner/behav", nh);
+		sm_behav_notifier->initialize();
+		sm_behav_planner = std::make_shared<LocalPlannerStateMachine>(sm_behav_notifier, guard_local_planner);
+		sm_behav_planner->setStartFunction(std::bind(&HotaruTebLocalPlannerNode::onStart,this));
 		sm_behav_planner->setCbRelay(std::bind(&HotaruTebLocalPlannerNode::onRelay, this));
 		sm_behav_planner->setCbReplanningEnter(std::bind(&HotaruTebLocalPlannerNode::onReplan, this));
 		sm_behav_planner->setCbWaiting(std::bind(&HotaruTebLocalPlannerNode::onWaiting, this));
@@ -382,7 +349,6 @@ public:
 			geometry_msgs::Point p;
 			geometry_msgs::Point p_Wro;
 			int _via_lookaheadpoint = TrajectoryLookaheadPoint::lookaheadPoint(1.5, pubsubstate->msg_port_current_velocity.twist);
-			// There are some points that are definetely not needed
 			for (
 				int i = pubsubstate->msg_port_closests_waypoint.data+_via_lookaheadpoint;
 					i < trajectoryslicestrategy->getSlicePoint(); i++)
@@ -402,42 +368,50 @@ public:
 
 	virtual bool plancycle()
 	{
-		if (trajectoryslicestrategy->getSlicePoint()<0) return false;
-		// TODO add valid tf as synchronization requirement
-		if (std::isnan(se2_plan_end.x()) ||
-				std::isnan(se2_plan_end.y()) ||
-				std::isnan(se2_plan_end.theta()))
+		if (sm_behav_planner->isReplanning()||pubsubstate->bypass_behavior)
 		{
-			ROS_ERROR("Invalid goal point selected");
+			if (trajectoryslicestrategy->getSlicePoint()<0) return false;
+			// TODO add valid tf as synchronization requirement
+			if (std::isnan(se2_plan_end.x()) ||
+					std::isnan(se2_plan_end.y()) ||
+					std::isnan(se2_plan_end.theta()))
+			{
+				ROS_ERROR("Invalid goal point selected");
+				return false;
+			}
+			mtx_planner.lock();
+			auto t_start = std::chrono::steady_clock::now();
+			bool plan_success = planner->plan(se2_current_pose, se2_plan_end);
+			auto t_end = std::chrono::steady_clock::now();
+			mtx_planner.unlock();
+			std::chrono::duration<double> elps = t_end-t_start;
+			if (plan_success)
+			{
+				mtx_trajectory_update.lock();
+				try
+				{
+					planner->getFullTrajectory(trajectory_profile);
+				}
+				catch(std::runtime_error &e)
+				{
+					ROS_ERROR("Replan error!");
+				}
+				mtx_trajectory_update.unlock();
+			}
+			if (pubsubstate->debug)
+			{
+				planner->visualize();
+				pubsubstate->msg_port_calc_planner_time.data = elps.count();
+				port_calc_planner_time.publish(pubsubstate->msg_port_calc_planner_time);
+			}
+			//std::cout << se2_current_pose << '\t' <<  se2_plan_end << '\t' << elps.count() << '\t' << trajectoryslicestrategy->getSlicePoint() << '\n';
+			return true;
+		}
+
+		else
+		{
 			return false;
 		}
-		mtx_planner.lock();
-		auto t_start = std::chrono::steady_clock::now();
-		bool plan_success = planner->plan(se2_current_pose, se2_plan_end);
-		auto t_end = std::chrono::steady_clock::now();
-		mtx_planner.unlock();
-		std::chrono::duration<double> elps = t_end-t_start;
-		if (plan_success)
-		{
-			mtx_trajectory_update.lock();
-			try
-			{
-				planner->getFullTrajectory(trajectory_profile);
-			}
-			catch(std::runtime_error &e)
-			{
-				ROS_ERROR("Replan error!");
-			}
-			mtx_trajectory_update.unlock();
-		}
-		if (pubsubstate->debug)
-		{
-			planner->visualize();
-			pubsubstate->msg_port_calc_planner_time.data = elps.count();
-			port_calc_planner_time.publish(pubsubstate->msg_port_calc_planner_time);
-		}
-		std::cout << se2_current_pose << '\t' <<  se2_plan_end << '\t' << elps.count() << '\t' << trajectoryslicestrategy->getSlicePoint() << '\n';
-		return true;
 	}
 
 };
