+++ hotaru/hotaru_planner_node_teb/launch/nissan-leaf/hotaru-nissan.leaf.rel.launch
@@ -1,0 +1,6 @@
+<launch>  
+  <include file="$(find hotaru_planner_node_teb)/launch/hotaru_autoware_interface.launch">
+  </include>
+  <include file="$(find nissan_sim_bringup)/launch/hotaru/test.localtrajectory.rel.launch">
+  </include>
+</launch>
