+++ rei/rei_monitoring_msgs/msg/DetectedObstacles.msg
@@ -1,2 +1,3 @@
 std_msgs/Header header
 rei_monitoring_msgs/Obstacle[] obstacles
+int32 max_closest_waypoint
