+++ rei/rei_monitoring_msgs/msg/Obstacle.msg
@@ -1,6 +1,7 @@
 geometry_msgs/Pose pose
 float64 radius
 geometry_msgs/Point[] convex_hull
+int32 closest_waypoint
 uint8 PERIMETER=1
 uint8 STATIC_OBSTACLE=1
 uint8 DYNAMIC_OBSTACLE=2
