+++ rei/rei_perception_monitor/CMakeLists.txt
@@ -125,7 +125,7 @@ catkin_package(
 ## Specify additional locations of header files
 ## Your package locations should be listed before other locations
 include_directories(
-# include
+  include
   ${catkin_INCLUDE_DIRS}
 )
 
@@ -142,7 +142,7 @@ include_directories(
 ## Declare a C++ executable
 ## With catkin_make all packages are built within a single CMake context
 ## The recommended prefix ensures that target names across packages don't collide
-add_executable(rei_obstacle_monitor_local_node src/obstacle_detector_node.cpp)
+add_executable(rei_obstacle_monitor_local_node src/obstacle_detector_node.cpp src/object_waypoint_detector.cpp)
 
 ## Rename C++ executable without prefix
 ## The above recommended prefix causes long target names, the following renames the
