+++ rei/rei_perception_monitor/scripts/test-perception/test-cases/test1.json
@@ -1,0 +1,16 @@
+
+{
+"local_planner_test":
+{
+	"obstacle": [ {
+		"x": 42.0639,
+		"y": -25.0102004,
+		"radius": "0.5"
+	},
+	{
+	    "x": 39.6821,
+		"y": -55.0188000007,
+		"radius": "0.5"
+	}]
+}
+}
